Add files via upload
The Sobel algorithm is commonly employed for edge detection, identifying boundaries where the gradient of the image is high. This project focuses on the VLSI implementation of the edge detection operation for iris images. Initially, the RGB image is converted to grayscale using MATLAB-Simulink HDL blocks. The grayscale image is then fed into Xilinx Vivado IDE, where Verilog HDL is utilized to apply the Sobel algorithm, extracting edge pixels.